stage 11 stage 6 make store folder
