+++ src/components/ButtonCartChangeCount/index.css
@@ -1,4 +1,4 @@
-.buttonCartIncrease {
+.buttonCartChangeCount {
   height: 2rem;
   width: 2rem;
   background-color: white;
