--- src/components/ButtonCartDecrease/index.css
@@ -1,8 +1,0 @@
-.buttonCartDecrease {
-  height: 2rem;
-  width: 2rem;
-  background-color: white;
-  margin: 0.5rem;
-  font-size: 0.8rem;
-  border: 1px solid black;
-}
