+++ src/containers/CartPage/index.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { connect } from 'react-redux';
 import { withRouter } from 'react-router';
 
 import HeaderCart from '../../components/HeaderCart';
