+++ src/store/App/selectors.js
@@ -1,0 +1,9 @@
+export const selectLoading = store => store.app.loading;
+
+export const selectError = store => store.app.error;
+
+export const selectSuccess = store => store.app.success;
+
+export const selectProductsData = store => store.app.productsData; //yet this function need later
+
+export const selectCategoriesList = store => store.app.categoriesList;
