+++ src/store/reducers.js
@@ -1,10 +1,10 @@
 import { combineReducers } from 'redux';
 import { connectRouter } from 'connected-react-router';
 
-import formfilter from './containers/FormFilter/reducers';
-import listcontainer from './containers/ListContainer/reducers';
-import app from './containers/App/reducers';
-import cart from './containers/Cart/reducers';
+import formfilter from './FormFilter/reducers';
+import listcontainer from './ListContainer/reducers';
+import app from './App/reducers';
+import cart from './Cart/reducers';
 
 const createRootReducer = history =>
 	combineReducers({
